resume_agent/requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
requests>=2.28.0
PyMuPDF>=1.23.0
pydantic>=2.12.5
sqlmodel>=0.0.14
openai>=1.12.0
python-dotenv>=1.0.0
